# Generate c signature when possible