Any reason to be conservative on using c signature before?

@vtjnash 
Fix #11304