@yuyichao 2015-05-17
The original commit that introduce this (https://github.com/JuliaLang/julia/commit/9a68a05fbb698d95f10b29170af4c207b13230c6) doesn't have a explaination for these checks.

@JeffBezanson