@yuyichao 2015-05-17
Test script:
```julia
@noinline g(a) = 1
@noinline f(a) = Int
@noinline f(a, b) = Int

@code_llvm g(Int)
@code_llvm f(Int)
@code_llvm f(Int, Int)

k1() = g(Int)
k2() = f(Int)
k3() = f(Int, Int)

@code_llvm k1()
@code_llvm k2()
@code_llvm k3()

function time_func(f::Function, args...)
    println(f)
    f(args...)
    gc()
    @time for i in 1:100000000
        f(args...)
    end
    gc()
end

time_func(k1)
time_func(k2)
time_func(k3)
```

Before:
```llvm
define i64 @julia_g_44381(%jl_value_t*) {
top:
  ret i64 1
}

define %jl_value_t* @julia_f_44384(%jl_value_t*, %jl_value_t**, i32) {
top:
  %3 = load %jl_value_t** inttoptr (i64 140245884996296 to %jl_value_t**), align 8
  ret %jl_value_t* %3
}

define %jl_value_t* @julia_f_44386(%jl_value_t*, %jl_value_t**, i32) {
top:
  %3 = load %jl_value_t** inttoptr (i64 140245884996296 to %jl_value_t**), align 8
  ret %jl_value_t* %3
}

define i64 @julia_k1_44390() {
top:
  %0 = load %jl_value_t** inttoptr (i64 140245884996296 to %jl_value_t**), align 8
  %1 = call i64 @julia_g_44381(%jl_value_t* %0)
  ret i64 %1
}

define %jl_value_t* @julia_k2_44391() {
top:
  %0 = alloca [3 x %jl_value_t*], align 8
  %.sub = getelementptr inbounds [3 x %jl_value_t*]* %0, i64 0, i64 0
  %1 = getelementptr [3 x %jl_value_t*]* %0, i64 0, i64 2
  %2 = bitcast [3 x %jl_value_t*]* %0 to i64*
  store i64 2, i64* %2, align 8
  %3 = getelementptr [3 x %jl_value_t*]* %0, i64 0, i64 1
  %4 = bitcast %jl_value_t** %3 to %jl_value_t***
  %5 = load %jl_value_t*** @jl_pgcstack, align 8
  store %jl_value_t** %5, %jl_value_t*** %4, align 8
  store %jl_value_t** %.sub, %jl_value_t*** @jl_pgcstack, align 8
  store %jl_value_t* null, %jl_value_t** %1, align 8
  %6 = load %jl_value_t** inttoptr (i64 140245884996296 to %jl_value_t**), align 8
  store %jl_value_t* %6, %jl_value_t** %1, align 8
  %7 = call %jl_value_t* @julia_f_44384(%jl_value_t* inttoptr (i64 140245917780336 to %jl_value_t*), %jl_value_t** %1, i32 1)
  %8 = load %jl_value_t*** %4, align 8
  store %jl_value_t** %8, %jl_value_t*** @jl_pgcstack, align 8
  ret %jl_value_t* %7
}

define %jl_value_t* @julia_k3_44392() {
top:
  %0 = alloca [4 x %jl_value_t*], align 8
  %.sub = getelementptr inbounds [4 x %jl_value_t*]* %0, i64 0, i64 0
  %1 = getelementptr [4 x %jl_value_t*]* %0, i64 0, i64 2
  %2 = bitcast [4 x %jl_value_t*]* %0 to i64*
  store i64 4, i64* %2, align 8
  %3 = getelementptr [4 x %jl_value_t*]* %0, i64 0, i64 1
  %4 = bitcast %jl_value_t** %3 to %jl_value_t***
  %5 = load %jl_value_t*** @jl_pgcstack, align 8
  store %jl_value_t** %5, %jl_value_t*** %4, align 8
  store %jl_value_t** %.sub, %jl_value_t*** @jl_pgcstack, align 8
  store %jl_value_t* null, %jl_value_t** %1, align 8
  %6 = getelementptr [4 x %jl_value_t*]* %0, i64 0, i64 3
  store %jl_value_t* null, %jl_value_t** %6, align 8
  %7 = load %jl_value_t** inttoptr (i64 140245884996296 to %jl_value_t**), align 8
  store %jl_value_t* %7, %jl_value_t** %1, align 8
  %8 = load %jl_value_t** inttoptr (i64 140245884996296 to %jl_value_t**), align 8
  store %jl_value_t* %8, %jl_value_t** %6, align 8
  %9 = call %jl_value_t* @julia_f_44386(%jl_value_t* inttoptr (i64 140245917882576 to %jl_value_t*), %jl_value_t** %1, i32 2)
  %10 = load %jl_value_t*** %4, align 8
  store %jl_value_t** %10, %jl_value_t*** @jl_pgcstack, align 8
  ret %jl_value_t* %9
}
k1
   1.298 seconds     
k2
   1.403 seconds     
k3
   1.465 seconds     
```

After
```llvm
define i64 @julia_g_67425(%jl_value_t*) {
top:
  ret i64 1
}

define %jl_value_t* @julia_f_67428(%jl_value_t*) {
top:
  %1 = load %jl_value_t** inttoptr (i64 139685242901352 to %jl_value_t**), align 8
  ret %jl_value_t* %1
}

define %jl_value_t* @julia_f_67430(%jl_value_t*, %jl_value_t*) {
top:
  %2 = load %jl_value_t** inttoptr (i64 139685242901352 to %jl_value_t**), align 8
  ret %jl_value_t* %2
}

define i64 @julia_k1_67434() {
top:
  %0 = load %jl_value_t** inttoptr (i64 139685242901352 to %jl_value_t**), align 8
  %1 = call i64 @julia_g_67425(%jl_value_t* %0)
  ret i64 %1
}

define %jl_value_t* @julia_k2_67435() {
top:
  %0 = load %jl_value_t** inttoptr (i64 139685242901352 to %jl_value_t**), align 8
  %1 = call %jl_value_t* @julia_f_67428(%jl_value_t* %0)
  ret %jl_value_t* %1
}

define %jl_value_t* @julia_k3_67436() {
top:
  %0 = load %jl_value_t** inttoptr (i64 139685242901352 to %jl_value_t**), align 8
  %1 = call %jl_value_t* @julia_f_67430(%jl_value_t* %0, %jl_value_t* %0)
  ret %jl_value_t* %1
}
k1
   1.288 seconds     
k2
   1.123 seconds     
k3
   1.109 seconds     
```

I don't really understand why is `k2` and `k3` even faster than `k1` now... (it's consistent and is not because of the order I run them...). Maybe sth like boxing of the result?