@yuyichao 2015-05-25
Ping.

Maybe any guideline on how much to specialize?

@JeffBezanson I noticed that https://github.com/JuliaLang/julia/commit/01c871c1b634156340b740df2883798a6266af4b enables more aggressive specialization and AFAIK has a even bigger impact (at least on the size of `sys.so`) than this one. What's the rule of thumb to determine how mush to specialize?