@JeffBezanson 2015-05-25
I think a good way to go here might be to allow signatures with 0, 1, 2, or 3 `jl_value_t*` arguments as part of the jlcall convention. In other words, we would need a tag along with every `jl_fptr_t` to tell us what arguments it takes. Then we wouldn't need so many jlcall wrappers.

Doing this well requires some thought. Ideally we would not literally add a tag to jl_fptr_t, since that would be a lot of extra indirection. We could instead use a lambda_info's cFunction directly, for example. In fact the entire jl_function_t type is kind of useless at this point, and needs a redesign. See discussion in #10269 for example.